+++ PowerShellForFOD/Public/Get-FODConfig.ps1
@@ -2,23 +2,16 @@ Function Get-FODConfig {
     <#
     .SYNOPSIS
         Get PowerShell For FOD module configuration.
-
     .DESCRIPTION
         Retrieves the PowerShell for FOD module configuration from the serialized XML file.
-
     .PARAMETER Source
         Get the config data from either:
-
             PS4FOD:     the live module variable used for command defaults
             PS4FOD.xml: the serialized PS4FOD.xml that loads when importing the module
-
         Defaults to PS4FOD.
-
     .PARAMETER Path
         If specified, read config from this XML file.
-
         Defaults to PS4FOD.xml in the user temp folder on Windows, or .ps4fod in the user's home directory on Linux/macOS.
-
     .FUNCTIONALITY
         Fortify on Demand.
     #>
@@ -33,16 +26,12 @@ Function Get-FODConfig {
         $Path = $script:_PS4FODXmlpath
     )
 
-    if($PSCmdlet.ParameterSetName -eq 'source' -and $Source -eq "PS4FOD" -and -not $PSBoundParameters.ContainsKey('Path'))
-    {
+    if($PSCmdlet.ParameterSetName -eq 'source' -and $Source -eq "PS4FOD" -and -not $PSBoundParameters.ContainsKey('Path')) {
         $Script:PS4FOD
-    }
-    else
-    {
+    } else {
         function Decrypt {
             param($String)
-            if($String -is [System.Security.SecureString])
-            {
+            if($String -is [System.Security.SecureString]) {
                 [System.Runtime.InteropServices.marshal]::PtrToStringAuto(
                         [System.Runtime.InteropServices.marshal]::SecureStringToBSTR(
                                 $string))